==> 01_poisson_reg.Rmd <==
---
title: "01_regression"
author: "Brenna Kelly"
date: "2024-02-09"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## setup

```{r}

library(sp)
library(sf)
library(tmap)
library(INLA)
library(spdep)
library(dplyr)
library(moments)
library(ggplot2)
library(stringr)
library(regclass)

```

## data

```{r}

# acs <- acs_merge

acs <- read.csv("data/7020_data_count.csv")

acs$geoid <- str_pad(acs$geoid, width = 12, pad = "0", side = "left")

head(acs)
# acs_geom <- st_read("data/7020_geom.shp")
# 
# # merge by geoid
# acs <- merge(acs_geom, acs, by = "geoid")

# drop empty geometries
# acs <- acs[which(st_is_empty(acs) == FALSE), ]
# 
# # drop areas with no population
# empty_areas <- acs[which(is.na(acs$hisp_p)), ]
# acs <- acs[which(!acs$geoid %in% empty_areas$geoid), ]

# fix poverty (some have 0)
acs$pov_p <- ifelse(is.na(acs$pov_p), 0 , acs$pov_p)

# scale 0-100
acs <- acs |>
  mutate(hisp_perc = ifelse(is.na(hisp_p), 0, hisp_p*100)) |>
  mutate(black_perc = ifelse(is.na(black_p), 0, black_p*100)) |>
  mutate(aian_perc = ifelse(is.na(aian_p), 0, aian_p*100)) |>
  mutate(asian_perc = ifelse(is.na(asian_p), 0, asian_p*100)) |>
  mutate(nhpi_perc = ifelse(is.na(nhpi_p), 0, nhpi_p*100)) |>
  mutate(other_perc = ifelse(is.na(other_p), 0, other_p*100)) |>
  mutate(tom_perc = ifelse(is.na(tom_p), 0, tom_p*100)) |>
  mutate(pov_perc = ifelse(is.na(pov_p), 0, pov_p*100))
acs <- acs |>
  mutate(hisp_p_log = log(hisp_perc + min(acs$hisp_perc[which(acs$hisp_perc > 0)]))) |>
  mutate(black_p_log = log(black_perc + min(acs$black_perc[which(acs$black_perc > 0)]))) |>
  mutate(aian_p_log = log(aian_perc + min(acs$aian_perc[which(acs$aian_perc > 0)]))) |>
  mutate(asian_p_log = log(asian_perc + min(acs$asian_perc[which(acs$asian_perc > 0)]))) |>
  mutate(nhpi_p_log = log(nhpi_perc + min(acs$nhpi_perc[which(acs$nhpi_perc > 0)]))) |>
  mutate(other_p_log = log(other_perc + min(acs$other_perc[which(acs$other_perc > 0)]))) |>
  mutate(tom_p_log = log(tom_perc + min(acs$tom_perc[which(acs$tom_perc > 0)]))) |>
  mutate(pov_p_log = log(pov_perc + min(acs$pov_perc[which(acs$pov_perc > 0)]))) |>
  mutate(total_1k = total / 1000) |>
  mutate(total_1k_log = log(total_1k + min(acs$total_1k[which(acs$total_1k > 0)])))

acs$state <- str_sub(acs$geoid, start = 0, end = 2)

acs$total_centered <- acs$total_1k - mean(acs$total_1k)
# acs$total_1k_log <- log(acs$total_1k)
skewness(acs$total_1k)
# write.csv(acs, "acs_inla.csv", row.names = FALSE)

```


```{r}

#sources <- read.csv("/Users/brenna/Downloads/Air Quality System (AIRS AQS).csv")

# icis_minor <- read.csv("/Users/brenna/Downloads/Integrated Compliance Information System-Air (ICIS-Air).csv")
icis <- read.csv("/Users/brenna/Downloads/ICIS-AIR_downloads/ICIS-AIR_FACILITIES.csv")

addresses <- icis[, c("REGISTRY_ID", "STREET_ADDRESS", "CITY", "STATE", "ZIP_CODE")]
addresses$file <- rep(c(1:(28*2)), length = nrow(addresses))
table(addresses$file)
addr_files <- split(addresses, addresses$file)

for(i in 1:length(addr_files)) {
  name <- paste0("address_file_", i, ".csv")
  write.table(addr_files[i], file = name, col.names = FALSE, row.names = FALSE, sep = ",")
}

write.csv()

fac <- read.csv("/Users/brenna/Downloads/national_combined/NATIONAL_FACILITY_FILE.CSV")
facilities <- fac[, c("REGISTRY_ID", "CENSUS_BLOCK_CODE", "LATITUDE83", "LONGITUDE83")]
head(fac)
table(is.na(facilities$CENSUS_BLOCK_CODE))


test <- merge(facilities, icis, by = "REGISTRY_ID")
test <- test |>
  filter(!is.na(LATITUDE83) & !is.na(LONGITUDE83) & !is.na(CENSUS_BLOCK_CODE))

test$CENSUS_BLOCK_CODE <- str_pad(test$CENSUS_BLOCK_CODE, width = 15, side = "left", pad = "0")
test$census_block_group <- str_sub(test$CENSUS_BLOCK_CODE, start = 1, end = 12)

n_sources <- test |>
  count(census_block_group)

# merge with acs
test_acs <- merge(acs, n_sources, by.x = "geoid", by.y = "census_block_group", all.x = TRUE)
prop.table(table(test_acs$n == 0))
test_acs$n <- ifelse(is.na(test_acs$n), 0, test_acs$n)
test_acs$n_test <- test_acs$n + 1
test_acs$n_test_log <- log(test_acs$n + 1)

test_acs$n_test <- ifelse(test_acs$n > 500, 1, test_acs$n)

skewness(test_acs$n / (test_acs$total_1k + 1))


```



```{r}
summary(test_acs)
test <- inla(n_lower_o3 ~ offset(total_1k),# + 
             #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
             #f(idarea_2, model = "iid"), 
             data = test_acs, family = "poisson",
             control.inla= list(int.strategy = "eb"),
             control.compute = list(dic = TRUE, waic = TRUE),
             control.predictor = list(compute = TRUE))

summary(test)
1.77e+218
round(exp(test$summary.fixed), 2)

1 - exp(-27/1621*10)



```


### scaling, transformations

```{r}

# skewness(log(acs$hisp_p + min(acs$hisp_p[which(acs$hisp_p > 0)])), na.rm = TRUE)
# 
# skewness(acs$hisp_p)



# summary(acs)

# skewness(log(acs$pov_p + min(acs$pov_p[which(acs$pov_p > 0)])), na.rm = TRUE)

```

## pre-analysis exploration

```{r}

aggregate(acs$black_p, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$black_p, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

aggregate(acs$aian_p, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$aian_p, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

aggregate(acs$hisp_p, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$hisp_p, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

aggregate(acs$pov_p, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$pov_p, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

aggregate(acs$asian_p, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$asian_p, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

aggregate(acs$tom_p, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$tom_p, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

aggregate(acs$other_p, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$other_p, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

aggregate(acs$total, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$total, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

aggregate(acs$coll_p, by = list(acs$o3_measured_u), FUN = mean, na.rm = TRUE)
aggregate(acs$coll_p, by = list(acs$o3_measured_l), FUN = mean, na.rm = TRUE)

```

## spatial structure

```{r}

# area id for each observation
# acs$idarea <- 1:nrow(acs)
# 
# ## building the spatial weight matrix
# 
# # centroids
# acs.geom <- st_geometry(acs)
# acs.coords <- st_centroid(acs.geom)
# 
# # plot(acs.geom, reset = FALSE)
# # plot(acs.coords, pch = 16, col = 2, add = TRUE)
# 
# # queen's case boundaries
# acs_nb <- poly2nb(acs)
# 
# summary(acs_nb)
# acs_nb
# # acs_nb <- acs_nb[-c(3928, 3929, 15237, 20928, 33392, 36909, 48512, 
# #                     53416, 53423, 53472, 96818, 97230, 97256, 97605, 
# #                     102598, 107669, 109842, 110179, 110183, 137369, 
# #                     141538, 153359, 155904, 160601, 167742, 172373, 
# #                     191215, 230418, 231154, 231156, 231362, 232786,
# #                     234668, 235531)]
# acs_nb
# 
# no_neighbors <- c(3928, 3929, 15237, 20928, 33392, 36909, 48512, 
#                   53416, 53423, 53472, 96818, 97230, 97256, 97605, 
#                   102598, 107669, 109842, 110179, 110183, 137369, 
#                   141538, 153359, 155904, 160601, 167742, 172373, 
#                   191215, 230418, 231154, 231156, 231362, 232786,
#                   234668, 235531)
# 
# no_neighbors <- acs[c(3928, 3929, 15237, 20928, 33392, 36909, 48512, 
#                     53416, 53423, 53472, 96818, 97230, 97256, 97605, 
#                     102598, 107669, 109842, 110179, 110183, 137369, 
#                     141538, 153359, 155904, 160601, 167742, 172373, 
#                     191215, 230418, 231154, 231156, 231362, 232786,
#                     234668, 235531), ]# |>
#   #st_drop_geometry()
# no_neighbors
# no_neighbors[1]
# 
# acs <- acs[-c(no_neighbors), ]
# 
# acs[c(no_neighbors), ]
# 
# acs_nb <- poly2nb(acs)
# summary(acs_nb)
# 
# nb_contiguity <- st_contiguity(acs)
# nb_k1 <- st_knn(st_centroid(acs), 1)
# nb_mixed <- nb_union(nb_contiguity, nb_k1)
# wt <- st_weights(nb_mixed)
# 
# test <- st_weights(nb_k1)
# 
# 
# acs_nb[ c(3928, 3929, 15237, 20928, 33392, 36909, 48512, 
#                   53416, 53423, 53472, 96818, 97230, 97256, 97605, 
#                   102598, 107669, 109842, 110179, 110183, 137369, 
#                   141538, 153359, 155904, 160601, 167742, 172373, 
#                   191215, 230418, 231154, 231156, 231362, 232786,
#                   234668, 235531)]
# 
# test <- knn2nb(knearneigh(acs.coords, k = 1))
# summary(test)
# 
# no_neighbors[1]
# 
# acs[]
# 
# acs_nb[[3928]]
# 
# acs_nb[[230418]]
# 
# acs[no_neighbors[1], ]
# 
# acs_nb[[1]]
# 
# st_nearest_feature(acs[no_neighbors[1], ], acs)
# 
# 
# Syr2 <- as_Spatial(acs)
# coords <- coordinates(Syr2)
# Sy2_nb <- edit.nb(acs_nb, coords, polys = Syr2)
# 
# edit(acs_nb, coords)
# #Sy2_nb <- edit.nb(acs_nb, acs.coords, polys = acs)
# 
# 
# c(no_neighbors[1] - 1, no_neighbors[1] + 1)
# 
# str(no_neighbors[1])
# geoid <- no_neighbors[1, "geoid"] |>
#   st_drop_geometry()
# acs_nb[14665]
# st_nearest_feature(no_neighbors[1, ], acs)
# 
# no_neighbors[i, ]
# # find the closest region, in terms of fips
# for(i in 1:length(no_neighbors)) {
#   
#   index <- no_neighbors[i]
#   
#   acs_nb[[index]] <- c(no_neighbors[1] - 1, no_neighbors[1] + 1)
#     #index - 1 #st_nearest_feature(acs[no_neighbors[i], ], acs)
#   
# }
# 
# 
# st_nearest_feature(acs[no_neighbors[235530], ], acs)
# str(acs_nb)
# 
# # # delauney triangulation
# # Sy3_nb <- tri2nb(Syracuse.coords)
# # 
# # # binary spatial weights (more neighbors not downweighted; i.e., style != 'W')
# # Sy1_lw_B <- nb2listw(Sy1_nb, style = 'B')
# # inverse distance weighting
# dists <- nbdists(acs_nb, acs.coords)
# 
# inverse_distance <- function(x) {1/(x/1000)}
# 
# idw <- lapply(dists, inverse_distance)
# 
# acs_lw_idwB <- nb2listw(acs_nb, glist = idw, style = "B", zero.policy = TRUE)

```

### global Moran's I test for autocorrelation

```{r}
# 
# moran.test(acs$pm_measured_l, 
#            listw = acs_lw_idwB, 
#            alternative = "two.sided", 
#            randomisation = TRUE, zero.policy = TRUE)

# # monte carlo method
# moran.mc(boston$logCMEDV, 
#          listw = acs_lw_idwB, 
#          nsim = 999, 
#          alternative = 'greater')

```

### local Moran's I; spatial pattern of autocorrelation

```{r}

# lm1 = localmoran(acs$, 
#                  listw = acs_lw_idwB, 
#                  alternative = "two.sided")
# 
# acs$pval.bin <- as.factor(ifelse(acs$pval < 0.05, "Significant", "Not-significant"))
# 
# tm_shape(boston) + 
#   tm_fill("pval.bin") +
#   tm_borders() +
#   tm_layout(main.title = "Local Moran's I (z-scores)",
#             main.title.size = 1,
#             legend.position = c("left", "bottom"))

```

### spatial weight matrix

```{r}

# acs_knn <- knn2nb(knearneigh(acs.coords, k = 2))
# ?knearneigh
# 
# ?edit.nb
# 
# acs_w <- nb2mat(nb_k1, zero.policy = TRUE)
# 
# acs.listw <- nb2listw(acs_knn, zero.policy = TRUE) # neighbors
# # acs_w <- nb2mat(acs.listw) # INLA
# 

```

### inla graph

```{r}

# map
# map <- as_Spatial(acs$geometry)
#codes <- rgdal::make_EPSG()
#codes[which(codes$code == "4326"), ]
# map <- spTransform(map,
#                    CRS("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +ellps=GRS80 +datum=NAD83"))
# 
# sapply(slot(map, "polygons"), function(x){slot(x, "ID")})
# 
# 
# map@data$geoid <- acs$geoid
# rownames(map@data) <- acs$geoid
# 
# rownames(acs) <- 1:nrow(acs)
# 
# SpatialPolygonsDataFrame(map, acs), match.ID = TRUE)
# 
# 
# nb <- poly2nb(map)

nb <- poly2nb(acs)
nb2INLA("map.adj", nb)
g <- inla.read.graph(filename = "map.adj")

acs$idarea <- 1:nrow(acs)
acs$idarea_2 <- 1:nrow(acs)

```



## regression with INLA
not looping because... ugh

```{r}

## running 12 inla models (one for each pollutant-scale) and saving them into a list; the saving the list to an rds file

inla_results_pois <- list()

inla_results_pois
saveRDS(inla_results_pois, file = "data/inla_results_pois.rds")
test <- readRDS("data/inla_results_pois.rds")



inla_results_pois[["inla_co_l"]] <- inla(n_lower_co ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                    data = acs, family = "poisson",
                                    control.inla= list(int.strategy = "eb"),
                                    control.compute = list(dic = TRUE, waic = TRUE),
                                    control.predictor = list(compute = TRUE))
inla_results_pois[["inla_co_u"]] <- inla(n_upper_co ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                    data = acs, family = "poisson",
                                    control.inla= list(int.strategy = "eb"),
                                    control.compute = list(dic = TRUE, waic = TRUE),
                                    control.predictor = list(compute = TRUE))

inla_results_pois[["inla_no2_l"]] <- inla(n_lower_no2 ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                     data = acs, family = "poisson",
                                     control.inla= list(int.strategy = "eb"),
                                     control.compute = list(dic = TRUE, waic = TRUE),
                                     control.predictor = list(compute = TRUE))
inla_results_pois[["inla_no2_u"]] <- inla(n_upper_no2 ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                     data = acs, family = "poisson",
                                     control.inla= list(int.strategy = "eb"),
                                     control.compute = list(dic = TRUE, waic = TRUE),
                                     control.predictor = list(compute = TRUE))

inla_results_pois[["inla_pm_l"]] <- inla(n_lower_pm ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"),   
                                    data = acs, family = "poisson",
                                    control.inla= list(int.strategy = "eb"),
                                    control.compute = list(dic = TRUE, waic = TRUE),
                                    control.predictor = list(compute = TRUE))
inla_results_pois[["inla_pm_u"]] <- inla(n_upper_pm ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                    data = acs, family = "poisson",
                                    control.inla= list(int.strategy = "eb"),
                                    control.compute = list(dic = TRUE, waic = TRUE),
                                    control.predictor = list(compute = TRUE))

inla_results_pois[["inla_pb_l"]] <- inla(n_lower_pb ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k_log,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                    data = acs, family = "poisson",
                                    control.inla= list(int.strategy = "eb"),
                                    control.compute = list(dic = TRUE, waic = TRUE),
                                    control.predictor = list(compute = TRUE))
inla_results_pois[["inla_pb_u"]] <- inla(n_upper_pb ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                    data = acs, family = "poisson",
                                    control.inla= list(int.strategy = "eb"),
                                    control.compute = list(dic = TRUE, waic = TRUE),
                                    control.predictor = list(compute = TRUE))

inla_results_pois[["inla_so2_l"]] <- inla(n_lower_so2 ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                     data = acs, family = "poisson",
                                     control.inla= list(int.strategy = "eb"),
                                     control.compute = list(dic = TRUE, waic = TRUE),
                                     control.predictor = list(compute = TRUE))
inla_results_pois[["inla_so2_u"]] <- inla(n_upper_so2 ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k,# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                     data = acs, family = "poisson",
                                     control.inla= list(int.strategy = "eb"),
                                     control.compute = list(dic = TRUE, waic = TRUE),
                                     control.predictor = list(compute = TRUE))

inla_results_pois[["inla_o3_l"]] <- inla(n_lower_o3 ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k_log,# + offset(total_1k),# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"),  
                                    data = acs, family = "poisson",
                                    control.inla= list(int.strategy = "eb"),
                                    control.compute = list(dic = TRUE, waic = TRUE),
                                    control.predictor = list(compute = TRUE))
inla_results_pois[["inla_o3_u"]] <- inla(n_upper_o3 ~ hisp_p_log + black_p_log + 
                                      aian_p_log + asian_p_log + nhpi_p_log + 
                                      other_p_log + tom_p_log + total_1k_log,# + offset(total_1k),# + 
                                      #f(idarea, model = "besag", graph = g_id, scale.model = TRUE) +
                                      #f(idarea_2, model = "iid"), 
                                    data = acs, family = "poisson",
                                    control.inla= list(int.strategy = "eb"),
                                    control.compute = list(dic = TRUE, waic = TRUE),
                                    control.predictor = list(compute = TRUE))

VIF(glm(n_upper_o3 ~ hisp_p_log + black_p_log + 
          aian_p_log + asian_p_log + nhpi_p_log + 
          other_p_log + tom_p_log + total_1k_log, data = acs, family = "poisson"))
# VIF is very low


```


```{r}

res_lower <- inla_results_pois$inla_o3_l$summary.fixed
res_upper <- inla_results_pois$inla_o3_u$summary.fixed
res_lower$label <- "lower"
res_upper$label <- "upper"

res <- rbind(res_lower, res_upper)
# res <- o3_res #summary(fit_intx)$coefs.SE.CI

res <- tibble::rownames_to_column(res, "variable")

res$variable <- ifelse(res$variable == "total_1k", "pop", res$variable)
res$variable <- ifelse(res$variable == "total_1k1", "pop1", res$variable)
res$label <- case_when(grepl("hisp", res$variable) == TRUE ~ "Hispanic or Latinx",
                       grepl("black", res$variable) == TRUE ~ "Black or African American",
                       grepl("aian", res$variable) == TRUE ~ "American Indian or Alaska Native",
                       grepl("asian", res$variable) == TRUE ~ "Asian",
                       grepl("nhpi", res$variable) == TRUE ~ "Native Hawaiian or Other Pacific Islander",
                       grepl("other", res$variable) == TRUE ~ "Some Other Race",
                       grepl("tom", res$variable) == TRUE ~ "Two or More Races",
                       grepl("pop", res$variable) == TRUE ~ "Population, 1K")
res$label <- ifelse(grepl("1", res$variable) == TRUE, paste0(res$label, ", upper"),
                    paste0(res$label, ", lower"))
res$scale <- ifelse(grepl("1", res$variable) == TRUE, "upper",
                    "lower")

res$mean <- round(exp(res$mean), 3)
res$`0.025quant` <- round(exp(res$`0.025quant`), 3)
res$`0.975quant` <- round(exp(res$`0.975quant`), 3)

res <- res[, c("variable", "mean", "0.025quant", "0.975quant", "scale", "label")]

res$sig <- ifelse(res$`0.025quant` < 1 & res$`0.025quant` < 1, "neg", NA)
res$sig <- ifelse(res$`0.975quant` > 1 & res$`0.975quant` > 1, "pos", res$sig)

# figure

res |>
  filter(!variable %in% c("(Intercept)", "(Intercept)1")) |>
  ggplot(aes(y = label, color = scale, group = scale)) +
  theme_classic() +
  geom_point(aes(x = mean), size = 1) +
  #scale_size_manual(values = c(1, 3)) +
  scale_color_manual(values = c("#ff2908","#f78e75")) +
  geom_linerange(aes(xmin = `0.025quant`, xmax = `0.975quant`)) +
  labs(x = "Odds of having O3 monitor (log scale)") +
  coord_cartesian(xlim = c(0.8, 1.1)) +
  geom_vline(xintercept = 1, linetype = "solid") +
  geom_vline(xintercept = c(0.8, 0.9, 1.1, 1.2), alpha = 0.5, linetype="dotted") +
  scale_x_continuous(trans = 'log10') +
  theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.y = element_blank()) + 
  geom_text(aes(x = (`0.975quant` + 0.02), y=label, label = round(mean, 2)), hjust = 0, fontface = "bold") +#, size = 4) +
  geom_text(aes(x = 0.8, y = label, label = stringr::str_wrap(label, 30), hjust=0), fontface = "plain", size = 4)

```


```{r pm}
res_lower <- inla_results_pois$inla_pm_l$summary.fixed
res_upper <- inla_results_pois$inla_pm_u$summary.fixed
res_lower$label <- "lower"
res_upper$label <- "upper"

res <- rbind(res_lower, res_upper)
# res <- o3_res #summary(fit_intx)$coefs.SE.CI

res <- tibble::rownames_to_column(res, "variable")

res$variable <- ifelse(res$variable == "total_1k", "pop", res$variable)
res$variable <- ifelse(res$variable == "total_1k1", "pop1", res$variable)
res$label <- case_when(grepl("hisp", res$variable) == TRUE ~ "Hispanic or Latinx",
                       grepl("black", res$variable) == TRUE ~ "Black or African American",
                       grepl("aian", res$variable) == TRUE ~ "American Indian or Alaska Native",
                       grepl("asian", res$variable) == TRUE ~ "Asian",
                       grepl("nhpi", res$variable) == TRUE ~ "Native Hawaiian or Other Pacific Islander",
                       grepl("other", res$variable) == TRUE ~ "Some Other Race",
                       grepl("tom", res$variable) == TRUE ~ "Two or More Races",
                       grepl("pop", res$variable) == TRUE ~ "Population, 1K")
res$label <- ifelse(grepl("1", res$variable) == TRUE, paste0(res$label, ", upper"),
                    paste0(res$label, ", lower"))
res$scale <- ifelse(grepl("1", res$variable) == TRUE, "upper",
                    "lower")


res$mean <- round(exp(res$mean), 3)
res$`0.025quant` <- round(exp(res$`0.025quant`), 3)
res$`0.975quant` <- round(exp(res$`0.975quant`), 3)

res <- res[, c("variable", "mean", "0.025quant", "0.975quant", "scale", "label")]

res$sig <- ifelse(res$`0.025quant` < 1 & res$`0.025quant` < 1, "neg", NA)
res$sig <- ifelse(res$`0.975quant` > 1 & res$`0.975quant` > 1, "pos", res$sig)

# figure

res |>
  filter(!variable %in% c("(Intercept)", "(Intercept)1")) |>
  ggplot(aes(y = label, color = scale, group = scale)) +
  theme_classic() +
  geom_point(aes(x = mean), size = 1) +
  #scale_size_manual(values = c(1, 3)) +
  scale_color_manual(values = c("#22b573", "#90d9b8")) +
  geom_linerange(aes(xmin = `0.025quant`, xmax = `0.975quant`)) +
  labs(x = "Odds of having O3 monitor (log scale)") +
  coord_cartesian(xlim = c(0.8, 1.1)) +
  geom_vline(xintercept = 1, linetype = "solid") +
  geom_vline(xintercept = c(0.8, 0.9, 1.1, 1.2), alpha = 0.5, linetype="dotted") +
  scale_x_continuous(trans = 'log10') +
  theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.y = element_blank()) + 
  geom_text(aes(x = (`0.975quant` + 0.02), y=label, label = round(mean, 2)), hjust = 0, fontface = "bold") +#, size = 4) +
  geom_text(aes(x = 0.8, y = label, label = stringr::str_wrap(label, 30), hjust=0), fontface = "plain", size = 4)

```

```{r no2}
res_lower <- inla_results_pois$inla_no2_l$summary.fixed
res_upper <- inla_results_pois$inla_no2_u$summary.fixed
res_lower$label <- "lower"
res_upper$label <- "upper"

res <- rbind(res_lower, res_upper)
# res <- o3_res #summary(fit_intx)$coefs.SE.CI

res <- tibble::rownames_to_column(res, "variable")

res$variable <- ifelse(res$variable == "total_1k", "pop", res$variable)
res$variable <- ifelse(res$variable == "total_1k1", "pop1", res$variable)
res$label <- case_when(grepl("hisp", res$variable) == TRUE ~ "Hispanic or Latinx",
                       grepl("black", res$variable) == TRUE ~ "Black or African American",
                       grepl("aian", res$variable) == TRUE ~ "American Indian or Alaska Native",
                       grepl("asian", res$variable) == TRUE ~ "Asian",
                       grepl("nhpi", res$variable) == TRUE ~ "Native Hawaiian or Other Pacific Islander",
                       grepl("other", res$variable) == TRUE ~ "Some Other Race",
                       grepl("tom", res$variable) == TRUE ~ "Two or More Races",
                       grepl("pop", res$variable) == TRUE ~ "Population, 1K")
res$label <- ifelse(grepl("1", res$variable) == TRUE, paste0(res$label, ", upper"),
                    paste0(res$label, ", lower"))
res$scale <- ifelse(grepl("1", res$variable) == TRUE, "upper",
                    "lower")

res$mean <- round(exp(res$mean), 3)
res$`0.025quant` <- round(exp(res$`0.025quant`), 3)
res$`0.975quant` <- round(exp(res$`0.975quant`), 3)

res <- res[, c("variable", "mean", "0.025quant", "0.975quant", "scale", "label")]

res$sig <- ifelse(res$`0.025quant` < 1 & res$`0.025quant` < 1, "neg", NA)
res$sig <- ifelse(res$`0.975quant` > 1 & res$`0.975quant` > 1, "pos", res$sig)

# figure

res |>
  filter(!variable %in% c("(Intercept)", "(Intercept)1")) |>
  ggplot(aes(y = label, color = scale, group = scale)) +
  theme_classic() +
  geom_point(aes(x = mean), size = 1) +
  #scale_size_manual(values = c(1, 3)) +
  scale_color_manual(values = c("#9e005d", "#ce7fad")) +
  geom_linerange(aes(xmin = `0.025quant`, xmax = `0.975quant`)) +
  labs(x = "Odds of having O3 monitor (log scale)") +
  coord_cartesian(xlim = c(0.8, 1.1)) +
  geom_vline(xintercept = 1, linetype = "solid") +
  geom_vline(xintercept = c(0.8, 0.9, 1.1, 1.2), alpha = 0.5, linetype="dotted") +
  scale_x_continuous(trans = 'log10') +
  theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.y = element_blank()) + 
  geom_text(aes(x = (`0.975quant` + 0.02), y=label, label = round(mean, 2)), hjust = 0, fontface = "bold") +#, size = 4) +
  geom_text(aes(x = 0.8, y = label, label = stringr::str_wrap(label, 30), hjust=0), fontface = "plain", size = 4)


```

```{r so2}
res_lower <- inla_results_pois$inla_so2_l$summary.fixed
res_upper <- inla_results_pois$inla_so2_u$summary.fixed
res_lower$label <- "lower"
res_upper$label <- "upper"

res <- rbind(res_lower, res_upper)
# res <- o3_res #summary(fit_intx)$coefs.SE.CI

res <- tibble::rownames_to_column(res, "variable")

res$variable <- ifelse(res$variable == "total_1k", "pop", res$variable)
res$variable <- ifelse(res$variable == "total_1k1", "pop1", res$variable)
res$label <- case_when(grepl("hisp", res$variable) == TRUE ~ "Hispanic or Latinx",
                       grepl("black", res$variable) == TRUE ~ "Black or African American",
                       grepl("aian", res$variable) == TRUE ~ "American Indian or Alaska Native",
                       grepl("asian", res$variable) == TRUE ~ "Asian",
                       grepl("nhpi", res$variable) == TRUE ~ "Native Hawaiian or Other Pacific Islander",
                       grepl("other", res$variable) == TRUE ~ "Some Other Race",
                       grepl("tom", res$variable) == TRUE ~ "Two or More Races",
                       grepl("pop", res$variable) == TRUE ~ "Population, 1K")
res$label <- ifelse(grepl("1", res$variable) == TRUE, paste0(res$label, ", upper"),
                    paste0(res$label, ", lower"))
res$scale <- ifelse(grepl("1", res$variable) == TRUE, "upper",
                    "lower")

res$mean <- round(exp(res$mean), 3)
res$`0.025quant` <- round(exp(res$`0.025quant`), 3)
res$`0.975quant` <- round(exp(res$`0.975quant`), 3)

res <- res[, c("variable", "mean", "0.025quant", "0.975quant", "scale", "label")]

res$sig <- ifelse(res$`0.025quant` < 1 & res$`0.025quant` < 1, "neg", NA)
res$sig <- ifelse(res$`0.975quant` > 1 & res$`0.975quant` > 1, "pos", res$sig)

# figure

res |>
  filter(!variable %in% c("(Intercept)", "(Intercept)1")) |>
  ggplot(aes(y = label, color = scale, group = scale)) +
  theme_classic() +
  geom_point(aes(x = mean), size = 1) +
  #scale_size_manual(values = c(1, 3)) +
  scale_color_manual(values = c("#0071bc", "#7fb7dd")) +
  geom_linerange(aes(xmin = `0.025quant`, xmax = `0.975quant`)) +
  labs(x = "Odds of having O3 monitor (log scale)") +
  coord_cartesian(xlim = c(0.8, 1.1)) +
  geom_vline(xintercept = 1, linetype = "solid") +
  geom_vline(xintercept = c(0.8, 0.9, 1.1, 1.2), alpha = 0.5, linetype="dotted") +
  scale_x_continuous(trans = 'log10') +
  theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.y = element_blank()) + 
  geom_text(aes(x = (`0.975quant` + 0.02), y=label, label = round(mean, 2)), hjust = 0, fontface = "bold") +#, size = 4) +
  geom_text(aes(x = 0.8, y = label, label = stringr::str_wrap(label, 30), hjust=0), fontface = "plain", size = 4)


```

```{r pb}

res_lower <- inla_results_pois$inla_pb_l$summary.fixed
res_upper <- inla_results_pois$inla_pb_u$summary.fixed
res_lower$label <- "lower"
res_upper$label <- "upper"

res <- rbind(res_lower, res_upper)
# res <- o3_res #summary(fit_intx)$coefs.SE.CI

res <- tibble::rownames_to_column(res, "variable")

res$variable <- ifelse(res$variable == "total_1k", "pop", res$variable)
res$variable <- ifelse(res$variable == "total_1k1", "pop1", res$variable)
res$label <- case_when(grepl("hisp", res$variable) == TRUE ~ "Hispanic or Latinx",
                       grepl("black", res$variable) == TRUE ~ "Black or African American",
                       grepl("aian", res$variable) == TRUE ~ "American Indian or Alaska Native",
                       grepl("asian", res$variable) == TRUE ~ "Asian",
                       grepl("nhpi", res$variable) == TRUE ~ "Native Hawaiian or Other Pacific Islander",
                       grepl("other", res$variable) == TRUE ~ "Some Other Race",
                       grepl("tom", res$variable) == TRUE ~ "Two or More Races",
                       grepl("pop", res$variable) == TRUE ~ "Population, 1K")
res$label <- ifelse(grepl("1", res$variable) == TRUE, paste0(res$label, ", upper"),
                    paste0(res$label, ", lower"))
res$scale <- ifelse(grepl("1", res$variable) == TRUE, "upper",
                    "lower")

res$mean <- round(exp(res$mean), 3)
res$`0.025quant` <- round(exp(res$`0.025quant`), 3)
res$`0.975quant` <- round(exp(res$`0.975quant`), 3)

res <- res[, c("variable", "mean", "0.025quant", "0.975quant", "scale", "label")]

res$sig <- ifelse(res$`0.025quant` < 1 & res$`0.025quant` < 1, "neg", NA)
res$sig <- ifelse(res$`0.975quant` > 1 & res$`0.975quant` > 1, "pos", res$sig)

# figure

res |>
  filter(!variable %in% c("(Intercept)", "(Intercept)1")) |>
  ggplot(aes(y = label, color = scale, group = scale)) +
  theme_classic() +
  geom_point(aes(x = mean), size = 1) +
  #scale_size_manual(values = c(1, 3)) +
  scale_color_manual(values = c("#c1272d", "#df9295")) +
  geom_linerange(aes(xmin = `0.025quant`, xmax = `0.975quant`)) +
  labs(x = "Odds of having O3 monitor (log scale)") +
  coord_cartesian(xlim = c(0.7, 1.4)) +
  geom_vline(xintercept = 1, linetype = "solid") +
  geom_vline(xintercept = c(0.8, 0.9, 1.1, 1.2), alpha = 0.5, linetype="dotted") +
  scale_x_continuous(trans = 'log10') +
  theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.y = element_blank()) + 
  geom_text(aes(x = (`0.975quant` + 0.02), y=label, label = round(mean, 2)), hjust = 0, fontface = "bold") +#, size = 4) +
  geom_text(aes(x = 0.7, y = label, label = stringr::str_wrap(label, 30), hjust=0), fontface = "plain", size = 4)


```

```{r co}
res_lower <- inla_results_pois$inla_co_l$summary.fixed
res_upper <- inla_results_pois$inla_co_u$summary.fixed
res_lower$label <- "lower"
res_upper$label <- "upper"

res <- rbind(res_lower, res_upper)
# res <- o3_res #summary(fit_intx)$coefs.SE.CI

res <- tibble::rownames_to_column(res, "variable")

res$variable <- ifelse(res$variable == "total_1k", "pop", res$variable)
res$variable <- ifelse(res$variable == "total_1k1", "pop1", res$variable)
res$label <- case_when(grepl("hisp", res$variable) == TRUE ~ "Hispanic or Latinx",
                       grepl("black", res$variable) == TRUE ~ "Black or African American",
                       grepl("aian", res$variable) == TRUE ~ "American Indian or Alaska Native",
                       grepl("asian", res$variable) == TRUE ~ "Asian",
                       grepl("nhpi", res$variable) == TRUE ~ "Native Hawaiian or Other Pacific Islander",
                       grepl("other", res$variable) == TRUE ~ "Some Other Race",
                       grepl("tom", res$variable) == TRUE ~ "Two or More Races",
                       grepl("pop", res$variable) == TRUE ~ "Population, 1K")
res$label <- ifelse(grepl("1", res$variable) == TRUE, paste0(res$label, ", upper"),
                    paste0(res$label, ", lower"))
res$scale <- ifelse(grepl("1", res$variable) == TRUE, "upper",
                    "lower")

res$mean <- round(exp(res$mean), 3)
res$`0.025quant` <- round(exp(res$`0.025quant`), 3)
res$`0.975quant` <- round(exp(res$`0.975quant`), 3)

res <- res[, c("variable", "mean", "0.025quant", "0.975quant", "scale", "label")]

res$sig <- ifelse(res$`0.025quant` < 1 & res$`0.025quant` < 1, "neg", NA)
res$sig <- ifelse(res$`0.975quant` > 1 & res$`0.975quant` > 1, "pos", res$sig)

# figure

res |>
  filter(!variable %in% c("(Intercept)", "(Intercept)1")) |>
  ggplot(aes(y = label, color = scale, group = scale)) +
  theme_classic() +
  geom_point(aes(x = mean), size = 1) +
  #scale_size_manual(values = c(1, 3)) +
  scale_color_manual(values = c("#00a99d", "#7fd3cd")) +
  geom_linerange(aes(xmin = `0.025quant`, xmax = `0.975quant`)) +
  labs(x = "Odds of having O3 monitor (log scale)") +
  coord_cartesian(xlim = c(0.8, 1.1)) +
  geom_vline(xintercept = 1, linetype = "solid") +
  geom_vline(xintercept = c(0.8, 0.9, 1.1, 1.2), alpha = 0.5, linetype="dotted") +
  scale_x_continuous(trans = 'log10') +
  theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.y = element_blank()) + 
  geom_text(aes(x = (`0.975quant` + 0.02), y=label, label = round(mean, 2)), hjust = 0, fontface = "bold") +#, size = 4) +
  geom_text(aes(x = 0.8, y = label, label = stringr::str_wrap(label, 30), hjust=0), fontface = "plain", size = 4)

```
